upd: split test func
